Move service_unavailable unit tests to test_error_responses.py
