+++ ocean_provider/utils/test/test_util.py
@@ -20,7 +20,6 @@ from ocean_provider.utils.util import (
     get_asset_urls,
     get_download_url,
     msg_hash,
-    service_unavailable,
 )
 from werkzeug.utils import get_content_type
 
@@ -32,53 +31,6 @@ def test_msg_hash():
     hashed = msg_hash(msg)
     expected = "7f83b1657ff1fc53b92dc18148a1d65dfc2d4b1fa3d677284addd200126d9069"
     assert hashed == expected
-
-
-def test_service_unavailable(caplog):
-    e = Exception("test message")
-    context = {"item1": "test1", "item2": "test2"}
-    response = service_unavailable(e, context, test_logger)
-    assert response.status_code == 503
-    response = response.json
-    assert response["error"] == "test message"
-    assert response["context"] == context
-    assert caplog.records[0].msg == "Payload was: item1=test1,item2=test2"
-
-
-def test_service_unavailable_strip_infura_project_id():
-    """Test that service_unavilable strips out infura project IDs."""
-
-    context = {"item1": "test1", "item2": "test2"}
-
-    # HTTP Infura URL (rinkeby)
-    e = Exception(
-        "429 Client Error: Too Many Requests for url: "
-        "https://rinkeby.infura.io/v3/ffffffffffffffffffffffffffffffff"
-    )
-    response = service_unavailable(e, context, test_logger)
-    assert (
-        response.json["error"] == "429 Client Error: Too Many Requests for url: "
-        "https://rinkeby.infura.io/v3/<infura project id stripped for security reasons>"
-    )
-
-    # Websocket Infura URL (ropsten)
-    e = Exception(
-        "429 Client Error: Too Many Requests for url: "
-        "https://ropsten.infura.io/ws/v3/ffffffffffffffffffffffffffffffff"
-    )
-    response = service_unavailable(e, context, test_logger)
-    assert (
-        response.json["error"] == "429 Client Error: Too Many Requests for url: "
-        "https://ropsten.infura.io/ws/v3/<infura project id stripped for security reasons>"
-    )
-
-    # Other
-    e = Exception("anything .infura.io/v3/<any_non_whitespace_string>")
-    response = service_unavailable(e, context, test_logger)
-    assert (
-        response.json["error"]
-        == "anything .infura.io/v3/<infura project id stripped for security reasons>"
-    )
 
 
 def test_build_download_response():
